test: split tests for theme toggle and area selectors functionality into separate files
